bangalore_traffic_agent/requirements.txt
google-adk
google-generativeai
crewai-tools
pydantic